./configure should look harder for GSS
This commit simplifies the search for a GSS library.

In principle we should only rely on our subset of the GSS-API C headers,
and dlopen()/dlsym()/dlclose().  The GSS-API C API does not have a
standard ABI, but it does have a de facto ABI, so we can rely on it.